Fix markdown heading anchor value
